backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
instagrapi>=2.0.0
apscheduler>=3.10.0
Pillow>=10.0.0
aiofiles>=23.2.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0